Chord constructors

These functions construct basic chord string notation from root notes.

Usage

chord_min(notes, key = "c", octaves = "tick")

chord_maj(notes, key = "c", octaves = "tick")

chord_min7(notes, key = "c", octaves = "tick")

chord_dom7(notes, key = "c", octaves = "tick")

chord_7s5(notes, key = "c", octaves = "tick")

chord_maj7(notes, key = "c", octaves = "tick")

chord_min6(notes, key = "c", octaves = "tick")

chord_maj6(notes, key = "c", octaves = "tick")

chord_dim(notes, key = "c", octaves = "tick")

chord_dim7(notes, key = "c", octaves = "tick")

chord_m7b5(notes, key = "c", octaves = "tick")

chord_aug(notes, key = "c", octaves = "tick")

chord_5(notes, key = "c", octaves = "tick")

chord_sus2(notes, key = "c", octaves = "tick")

chord_sus4(notes, key = "c", octaves = "tick")

chord_dom9(notes, key = "c", octaves = "tick")

chord_7s9(notes, key = "c", octaves = "tick")

chord_maj9(notes, key = "c", octaves = "tick")

chord_add9(notes, key = "c", octaves = "tick")

chord_min9(notes, key = "c", octaves = "tick")

chord_madd9(notes, key = "c", octaves = "tick")

chord_min11(notes, key = "c", octaves = "tick")

chord_7s11(notes, key = "c", octaves = "tick")

chord_maj7s11(notes, key = "c", octaves = "tick")

chord_11(notes, key = "c", octaves = "tick")

chord_maj11(notes, key = "c", octaves = "tick")

chord_13(notes, key = "c", octaves = "tick")

chord_min13(notes, key = "c", octaves = "tick")

chord_maj13(notes, key = "c", octaves = "tick")

xm(notes, key = "c", octaves = "tick")

xM(notes, key = "c", octaves = "tick")

xm7(notes, key = "c", octaves = "tick")

x7(notes, key = "c", octaves = "tick")

x7s5(notes, key = "c", octaves = "tick")

xM7(notes, key = "c", octaves = "tick")

xm6(notes, key = "c", octaves = "tick")

xM6(notes, key = "c", octaves = "tick")

xdim(notes, key = "c", octaves = "tick")

xdim7(notes, key = "c", octaves = "tick")

xm7b5(notes, key = "c", octaves = "tick")

xaug(notes, key = "c", octaves = "tick")

x5(notes, key = "c", octaves = "tick")

xs2(notes, key = "c", octaves = "tick")

xs4(notes, key = "c", octaves = "tick")

x9(notes, key = "c", octaves = "tick")

x7s9(notes, key = "c", octaves = "tick")

xM9(notes, key = "c", octaves = "tick")

xadd9(notes, key = "c", octaves = "tick")

xm9(notes, key = "c", octaves = "tick")

xma9(notes, key = "c", octaves = "tick")

xm11(notes, key = "c", octaves = "tick")

x7s11(notes, key = "c", octaves = "tick")

xM7s11(notes, key = "c", octaves = "tick")

x_11(notes, key = "c", octaves = "tick")

xM11(notes, key = "c", octaves = "tick")

x_13(notes, key = "c", octaves = "tick")

xm13(notes, key = "c", octaves = "tick")

xM13(notes, key = "c", octaves = "tick")

Arguments

notes

character, a noteworthy string of chord root notes.

key

key signature. See details.

octaves

character, passed to transpose().

Value

character

Details

Providing a key signature is used only to ensure flats or sharps for accidentals. An additional set of aliases with efficient names, of the form x* where * is a chord modifier abbreviation, is provided to complement the set of chord_* functions.

These functions create standard chords, not the multi-octave spanning types of chords commonly played on guitar.

See also

Examples

chord_min("d")
#> <Noteworthy string>
#>   Format: space-delimited time
#>   Values: <dfa>
chord_maj("d")
#> <Noteworthy string>
#>   Format: space-delimited time
#>   Values: <dg_a>
xM("d")
#> <Noteworthy string>
#>   Format: space-delimited time
#>   Values: <dg_a>
xm("c f g")
#> <Noteworthy string>
#>   Format: space-delimited time
#>   Values: <ce_g> <fa_c'> <gb_d'>
xm("c, f, g,", key = "e_")
#> <Noteworthy string>
#>   Format: space-delimited time
#>   Values: <c,e_,g,> <f,a_,c> <g,b_,d>