Save score to LilyPond file

Write score to a LilyPond format (.ly) text file for later use by LilyPond or subsequent editing outside of R.

lilypond(score, file, key = "c", time = "4/4", tempo = "2 = 60",
  header = NULL, string_names = NULL, paper = NULL, endbar = TRUE,
  midi = TRUE, out_dir = ".")

Arguments

score
a score object.

file
LilyPond output file, should end in .ly.

key
character, key signature, e.g., c, b_, f#m, etc.

time
character, defaults to "4/4".

tempo
character, defaults to "2 = 60".

header
a named list of arguments passed to the header of the LilyPond file. See details.

string_names
label strings at beginning of tab staff. NULL (default) for non-standard tunings only, TRUE or FALSE for force on or off completely.

paper
a named list of arguments for the LilyPond file page layout. See details.

endbar
character, the end bar.

midi
logical, add midi inclusion specification to LilyPond file.

out_dir
character, output directory, defaults to working directory.

Value

nothing returned; a file is written.

Details

The options for header include title, subtitle, composer, album, arranger, instrument, meter, opus, piece, poet, copyright and tagline. All header arguments are character strings. The paper argument list options are textheight, linewidth, indent, first_page_number, page_numbers and fontsize. All paper arguments are numeric except page_numbers, which is logical.

Examples

# not run