Note and pitch helpers

Helper functions for manipulating individual note and pitch strings.

note_is_natural(notes)

note_is_accidental(notes)

note_is_flat(notes)

note_is_sharp(notes)

naturalize(notes, type = c("both", "flat", "sharp"),
  ignore_octave = FALSE)

sharpen_flat(notes, ignore_octave = FALSE)

flatten_sharp(notes, ignore_octave = FALSE)

note_set_key(notes, key = "c")

note_rotate(notes, n = 0)

note_shift(notes, n = 0)

note_arpeggiate(notes, n = 0, ...)

pretty_notes(notes, ignore_octave = TRUE)

Arguments

notes
character, a noteworthy string, space-delimited or vector of individual entries.

type
character, type of note to naturalize.

ignore_octave
logical, strip any octave notation that may be present, returning only the basic notes without explicit pitch.

key
character, key signature to coerce any accidentals to the appropriate form for the key. May also specify "sharp" or "flat".

n
integer, degree of rotation.

...
additional arguments to transpose, specifically key and style.

Value

character

Details

In this context, sharpening flats and flattening sharps refers to inverting their notation, not raising and lowering a flatted or sharped note by one semitone. For the latter, use naturalize, which removes flat and/or sharp notation from a string.

Due to its simplicity, for note_rotate the strings may include chords. It simply rotates anything space-delimited or vectorized in place. Octave numbering is ignored if present.

By contrast, for note_shift the entire sequence is shifted up or down, as if inverting a broken chord. In this case notes is strictly interpreted and may not include chords. Octave numbering applies, though large, multi-octave gaps will be condensed in the process. Given the context of note_shift, the notes sequence should be ordered by increasing pitch. If it is not, ordering will be forced with each inversion during the n shifts.

note_arpeggiate also allows notes only. It is similar to note_shift, except that instead of a moving window, it grows from the original set of notes by n in the direction of the sign of n.

Examples

x <- "a_ a a#"
note_is_natural(x)

#> [1] FALSE  TRUE FALSE

note_is_accidental(x)

#> [1]  TRUE FALSE  TRUE

note_is_flat(x)

#> [1]  TRUE FALSE FALSE

note_is_sharp(x)

#> [1] FALSE FALSE  TRUE

note_set_key(x, "f")

#> <Noteworthy string>
#>   Format: space-delimited time
#>   Values: a_ a b_

note_set_key(x, "g")

#> <Noteworthy string>
#>   Format: space-delimited time
#>   Values: g# a a#

x <- "e_2 a_, c#f#a#"
naturalize(x)

#> <Noteworthy string>
#>   Format: space-delimited time
#>   Values: e2 a, <cfa>

naturalize(x, "sharp")

#> <Noteworthy string>
#>   Format: space-delimited time
#>   Values: e_2 a_, <cfa>

sharpen_flat(x)

#> <Noteworthy string>
#>   Format: space-delimited time
#>   Values: d#, g#, <c#f#a#>

flatten_sharp(x)

#> <Noteworthy string>
#>   Format: space-delimited time
#>   Values: e_, a_, <d_g_b_>

pretty_notes(x)

#> [1] "Eb Ab C#F#A#"

note_rotate(x, 1)

#> <Noteworthy string>
#>   Format: space-delimited time
#>   Values: a_, <c#f#a#> e_2

note_shift("c e g", 1)

#> <Noteworthy string>
#>   Format: space-delimited time
#>   Values: e g c4

note_shift("c e g", -4)

#> <Noteworthy string>
#>   Format: space-delimited time
#>   Values: g1 c2 e2

note_arpeggiate("c e g", 5)

#> <Noteworthy string>
#>   Format: space-delimited time
#>   Values: c e g c4 e4 g4 c5 e5

note_arpeggiate("c e g", -5)

#> <Noteworthy string>
#>   Format: space-delimited time
#>   Values: e1 g1 c2 e2 g2 c e g