Create a music score
Source: R/tracks.R

Create a music score from a collection of tracks.

score(track, chords = NULL, chord_seq = NULL)

Arguments

track
a track table consisting of one or more tracks.

chords
an optional named list of chords and respective fingerings generated by chord_set, for inclusion of a top center chord diagram chart.

chord_seq
an optional named vector of chords and their durations, for placing chord diagrams above staves in time.

Value

a score table.

Details

Score takes track tables generated by track and fortifies them as a music score. It optionally binds tracks with a set of chord diagrams. There may be only one track in track as well as no chord information passed, but for consistency score is still required to fortify the single track as a score object that can be rendered by tab.

Examples

x <- phrase("c ec'g' ec'g'", "4 4 2", "5 432 432")
x <- track(x)
score(x)

#> # A tibble: 1 x 7
#>   phrase       tuning         voice staff    ms_transpose ms_key tabstaff
#>   <list>       <chr>          <int> <chr>           <int> <chr>     <int>
#> 1 <S3: phrase> e, a, d g b e'     1 treble_8            0 <NA>          1