Repeat for tabr classes

Several methods are implemented for the classes noteworthy, noteinfo, and music. See tabr-methods for more details on methods for tabr classes.

# S3 method for noteworthy
rep(x, ...)

# S3 method for noteinfo
rep(x, ...)

# S3 method for music
rep(x, ...)

# S3 method for lyrics
rep(x, ...)

# S3 method for phrase
rep(x, ...)

Arguments

x
object.

...
additional arguments. Not accepted for phrase objects.

See also

Examples

# noteworthy class examples
x <- "a b c"
rep(x, 2)

#> [1] "a b c" "a b c"

rep(as_noteworthy(x), 2)

#> <Noteworthy string>
#>   Format: space-delimited time
#>   Values: a b c a b c

# noteinfo class examples
x <- "4x 4-.*2 2"
rep(as_noteinfo(x), times = c(2, 1, 1, 2))

#> <Note info string>
#>   Format: space-delimited time
#>   Values: 4x 4x 4-. 4-. 2 2

# music class examples
x <- "c,~4 c,1 c'e_'g'4-."
rep(as_music(x), each = 2)

#> <Music string>
#>   Format: space-delimited time
#>   Values: c,~4 c,~4 c,1 c,1 <c'e_'g'>4-. <c'e_'g'>4-.

# phrase class examples
rep(phrase(x), 2)

#> <Musical phrase>
#> <c,~>4 <c,>1 <c' es' g'>4-. <c,~>4 <c,>1 <c' es' g'>4-.