Options
Source: R/options.R

Options for tabr package.

tabr_options(...)

Arguments

...
a list of options.

Value

The function prints all set options if called with no arguments. When setting options, nothing is returned.

Details

Currently only lilypond, midi2ly and python are used. On Windows systems, if the system path for lilypond.exe, midi2ly.py and python.exe are not stored in the system PATH environmental variable, they must be provided by the user after loading the package.

Examples

tabr_options()

#> $dev
#> [1] "pdf"
#> 
#> $midi
#> [1] TRUE
#> 
#> $lilypond
#> [1] "C:/Program Files (x86)/LilyPond/usr/bin/lilypond.exe"
#> 
#> $midi2ly
#> [1] "C:/Program Files (x86)/LilyPond/usr/bin/midi2ly.py"
#> 
#> $python
#> [1] "C:/Program Files (x86)/LilyPond/usr/bin/python.exe"
#> 

tabr_options(lilypond = "C:/Program Files (x86)/LilyPond/usr/bin/lilypond.exe")