Create a volta/repeat phrase

Create a repeat section in LilyPond readable format.

volta(phrase, n = 1, endings = NULL, silent = FALSE)

pct(phrase, n = 1)

Arguments

phrase
a phrase or basic string to be repeated.

n
integer, number of repeats of phrase (one less than the total number of plays).

endings
list of phrases or basic strings, alternate endings.

silent
if TRUE, no text will be printed above the staff to indicate the number of repeats when greater than one and no endings = NULL. This is useful when another staff already displays the text.

Value

a phrase.

Details

This function takes a musical phrase or even a basic string and converts it into a volta phrase for the LilyPond volta engraver.

Note that basic strings should still be interpretable as a valid musical phrase by LilyPond. For example, a one-measure rest, "r1" does not need to be wrapped in a phrase class to work, nor would any other string explicitly written out in valid LilyPond syntax. As always, see the LilyPond documentation if you are not familiar with LilyPond syntax.

Examples

x <- phrase("c ec'g' ec'g'", "4 4 2", "5 432 432")
e1 <- phrase("a", 1, 5)
e2 <- phrase("b", 1, 5)
volta(x) # one repeat or two plays

#> \repeat volta 1 { <c\5>4 <e\4 c'\3 g'\2>4 <e\4 c'\3 g'\2>2 | }

pct(x) # as above, simple percent repeat notation

#> \repeat percent 2 { <c\5>4 <e\4 c'\3 g'\2>4 <e\4 c'\3 g'\2>2  }

volta(x, 1, list(e1, e2)) # one repeat with alternate ending

#> \repeat volta 1 { <c\5>4 <e\4 c'\3 g'\2>4 <e\4 c'\3 g'\2>2 | }
#> \alternative {
#>   { <a\5>1 | }
#>   { <b\5>1 | }
#> }

volta(x, 4, list(e1, e2)) # multiple repeats but with only one alternate ending

#> \repeat volta 4 { <c\5>4 <e\4 c'\3 g'\2>4 <e\4 c'\3 g'\2>2 | }
#> \alternative {
#>   { <a\5>1 | }
#>   { <b\5>1 | }
#> }