Append and duplicate

Helper functions for appending or pasting musical phrases and other strings together as well as repetition. The functions glue and dup are based on base functions paste andrep, respectively, but are tailored for efficiency in creating musical phrases. These functions respect and retain the phrase class when applied to phrases. Combining a phrase with a non-phrase string will assume compatibility and result in a new phrase object.

glue(...)

dup(x, n = 1)

Arguments

...
character, phrase or non-phrase string.

x
character, phrase or non-phrase string.

n
integer, number of repetitions.

Value

phrase on non-phrase character string.

Details

This is especially useful for repeated instances. This function applies to general slur notation as well. Multiple input formats are allowed. Total number of note durations must be even because all slurs require start and stop points.

Examples

glue(8, "16-", "8^")

#> [1] "8 16- 8^"

dup(1, 2)

#> [1] "1 1"

x <- phrase("c ec'g' ec'g'", "4 4 2", "5 432 432")
y <- phrase("a", 1, 5)
glue(x, y)

#> <c\5>4 <e\4 c'\3 g'\2>4 <e\4 c'\3 g'\2>2 <a\5>1

glue(x, dup(y, 2))

#> <c\5>4 <e\4 c'\3 g'\2>4 <e\4 c'\3 g'\2>2 <a\5>1 <a\5>1

glue(x, "r1") # add a simple rest instance

#> <c\5>4 <e\4 c'\3 g'\2>4 <e\4 c'\3 g'\2>2 r1

class(glue(x, y))

#> [1] "phrase"    "character"

class(dup(y, 2))

#> [1] "phrase"    "character"

class(glue(x, "r1"))

#> [1] "phrase"    "character"

class(dup("r1", 2))

#> [1] "character"

class(glue("r1", "r4"))

#> [1] "character"