Chord mapping

Helper functions for chord mapping.

gc_info(name, root_fret = NA, min_fret = NA, bass_string = NA,
  open = NA, key = "c", ignore_octave = FALSE)

gc_fretboard(name, root_fret = NA, min_fret = NA, bass_string = NA,
  open = NA, key = "c", ignore_octave = FALSE)

gc_notes(name, root_fret = NA, min_fret = NA, bass_string = NA,
  open = NA, key = "c", ignore_octave = FALSE)

chord_is_known(notes)

chord_name_split(name)

chord_name_root(name)

chord_name_mod(name)

Arguments

name
character, chord name in tabr format, e.g., "bM b_m b_m7#5", etc.

root_fret
integer, optional filter for chords whose root note matches a specific fret. May be a vector.

min_fret
integer, optional filter for chords whose notes are all at or above a specific fret. May be a vector.

bass_string
integer, optional filter for chords whose lowest pitch string matches a specific string, 6, 5, or 4. May be a vector.

open
logical, optional filter for open and movable chords. NA retains both types.

key
character, key signature, used to enforce type of accidentals.

ignore_octave
logical, if TRUE, functions like gc_info and gc_fretboard return more results.

notes
character, a noteworthy string.

Value

various, see details regarding each function.

Details

These functions assist with mapping between different information that define chords.

For chord_is_known, a check is done against chords in the guitarChords dataset. A simple noteworthy string is permitted, but any single-note entry will automatically yield a FALSE result.

gc_info returns a tibble data frame containing complete information for the subset of predefined guitar chords specified by name and key. Any accidentals present in the chord root of name (but not in the chord modifier, e.g., m7_5 or m7#5) are converted according to key if necessary. gc_notes and gc_fretboard are wrappers around gc_info, which return noteworthy strings of chord notes and a named vector of LilyPond fretboard diagram data, respectively. Note that although the input to these functions can contain multiple chord names, whether as a vector or as a single space-delimited string, the result is not intended to be of equal length. These functions filter guitarChords. The result is the set of all chords matched by the supplied input filters.

chord_name_split splits a vector or space-delimited set of chord names into a tibble data frame containing separate chord root and chord modifier columns. chord_name_root and chord_name_mod are simple wrappers around this.

Examples

chord_is_known("a b_,fb_d'f'")

#> [1] FALSE  TRUE

chord_name_root("a aM b_,m7#5")

#> [1] "a"   "a"   "b_,"

chord_name_mod("a aM b_,m7#5")

#> [1] "M"    "M"    "m7#5"

gc_info("a") # a major chord, not a single note

#> # A tibble: 3 x 12
#>   id    lp_name root  octave root_fret min_fret bass_string notes frets
#>   <fct> <chr>   <chr>  <dbl>     <dbl>    <dbl>       <int> <chr> <chr>
#> 1 M     a:5     a          3         7        7           4 ae'a~ xx79~
#> 2 M     a:5     a          3        12       12           5 ae'a~ x(12~
#> 3 M     a:5     a          3        12        9           5 ac#'~ x(12~
#> # ... with 3 more variables: semitones <list>, fretboard <chr>, open <lgl>

gc_info("ceg a#m7_5") # only third entry is a guitar chord

#> # A tibble: 3 x 12
#>   id    lp_name root  octave root_fret min_fret bass_string notes frets
#>   <fct> <chr>   <chr>  <dbl>     <dbl>    <dbl>       <int> <chr> <chr>
#> 1 m7_5  a#:m7_5 a#         3         8        4           4 a#c#~ xx86~
#> 2 m7_5  a#:m7_5 a#         3         8        8           4 a#e'~ xx89~
#> 3 m7_5  a#:m7_5 a#         3        13       12           5 a#g#~ x(13~
#> # ... with 3 more variables: semitones <list>, fretboard <chr>, open <lgl>

gc_info("ceg a#m7_5", key = "f")

#> # A tibble: 3 x 12
#>   id    lp_name root  octave root_fret min_fret bass_string notes frets
#>   <fct> <chr>   <chr>  <dbl>     <dbl>    <dbl>       <int> <chr> <chr>
#> 1 m7_5  b_:m7_5 b_         3         8        4           4 b_d_~ xx86~
#> 2 m7_5  b_:m7_5 b_         3         8        8           4 b_e'~ xx89~
#> 3 m7_5  b_:m7_5 b_         3        13       12           5 b_a_~ x(13~
#> # ... with 3 more variables: semitones <list>, fretboard <chr>, open <lgl>

gc_info("a,m c d f,")

#> # A tibble: 13 x 12
#>    id    lp_name root  octave root_fret min_fret bass_string notes frets
#>    <fct> <chr>   <chr>  <dbl>     <dbl>    <dbl>       <int> <chr> <chr>
#>  1 m     a,:m    a          2         0        0           5 a,ea~ xo22~
#>  2 m     a,:m    a          2         5        5           6 a,ea~ 5775~
#>  3 m     a,:m    a          2         5        2           6 a,cea 5322~
#>  4 M     c:5     c          3         3        3           5 cgc'~ x355~
#>  5 M     c:5     c          3         3        0           5 cegc~ x32o~
#>  6 M     c:5     c          3         8        8           6 cgc'~ 8(10~
#>  7 M     c:5     c          3         8        5           6 cegc~ 8755~
#>  8 M     d:5     d          3         0        0           4 dad'~ xxo2~
#>  9 M     d:5     d          3         5        5           5 dad'~ x577~
#> 10 M     d:5     d          3         5        2           5 df#a~ x542~
#> 11 M     d:5     d          3        10       10           6 dad'~ (10)~
#> 12 M     d:5     d          3        10        7           6 df#a~ (10)~
#> 13 M     f,:5    f          2         1        1           6 f,cf~ 1332~
#> # ... with 3 more variables: semitones <list>, fretboard <chr>, open <lgl>

gc_fretboard("a,m c d f,", 0:3)

#>           a,:m            c:5            c:5            d:5           f,:5 
#> "x;o;2;2;1;o;" "x;3;5;5;5;3;" "x;3;2;o;1;o;" "x;x;o;2;3;2;" "1;3;3;2;1;1;" 

x <- gc_notes("a, b,", 0:2)
summary(x)

#> <Noteworthy string>
#>   Timesteps: 2 (0 notes, 2 chords)
#>   Octaves: tick
#>   Accidentals: sharp
#>   Format: space-delimited time
#>   Values: <a,eac#'e'> <b,f#bd#'f#'>