LilyPond chord notation

Obtain LilyPond quasi-chord notation.

lp_chord_id(root, chord, exact = FALSE, ...)

lp_chord_mod(root, chord, exact = FALSE, ...)

Arguments

root

character, root note.

chord

character, tabr format chord name.

exact

logical, return a more exact LilyPond chord representation.

...

additional arguments passed to transpose().

Value

character

Details

These functions take a tabr syntax representation of a chord name and convert it to quasi-LilyPond syntax; "quasi" because the result still uses _ for flats and # for sharps, whereas LilyPond itself uses es and is (mostly). This is the format used by tabr functions involved in communicating with LilyPond for music transcription, and they make these final conversions on the fly. This can be overridden with exact = TRUE.

Examples

lp_chord_id("a a a", "m M m7_5")
#> [1] "a:m"    "a:5"    "a:m7_5"
lp_chord_mod("a a a", "m M m7_5")
#> [1] "m"    "5"    "m7_5"
lp_chord_id("a a a", "m M m7_5", exact = TRUE)
#> [1] "a:m"     "a:5"     "a:m7es5"
lp_chord_mod("a a a", "m M m7_5", exact = TRUE)
#> [1] "m"     "5"     "m7es5"