Note info helpers

Functions for working with note info strings.

Usage

info_duration(x)

info_slur_on(x)

info_slur_off(x)

info_slide(x)

info_bend(x)

info_dotted(x)

info_single_dotted(x)

info_double_dotted(x)

info_annotation(x)

info_articulation(x)

Arguments

x

character, note info string normally accompanying a noteworthy string for building phrase objects. x may also be a phrase object.

Value

character

Details

If x is a phrase object, there are some parsing limitations such as tuplets and repeats.

See also

Examples

a <- notate("t8x", "Start here")
notes <- "a, b, c d e f g# a r ac'e' a c' e' c' r*3 ac'e'~ ac'e'"
info <- paste(a, "t8x t8-. 16 4.. 16- 16 2^ 2 4. 8( 4)( 4) 8*4 1 1")
x <- as_music(notes, info)

data.frame(
  duration = info_duration(x),
  slur_on = info_slur_on(x),
  slur_off = info_slur_off(x),
  slide = info_slide(x),
  bend = info_bend(x),
  dotted = info_dotted(x),
  dotted1 = info_single_dotted(x),
  dotted2 = info_double_dotted(x),
  annotation = info_annotation(x),
  articulation = info_articulation(x)
)
#>    duration slur_on slur_off slide  bend dotted dotted1 dotted2 annotation
#> 1        t8   FALSE    FALSE FALSE FALSE  FALSE   FALSE   FALSE Start here
#> 2        t8   FALSE    FALSE FALSE FALSE  FALSE   FALSE   FALSE       <NA>
#> 3        t8   FALSE    FALSE FALSE FALSE  FALSE   FALSE   FALSE       <NA>
#> 4        16   FALSE    FALSE FALSE FALSE  FALSE   FALSE   FALSE       <NA>
#> 5       4..   FALSE    FALSE FALSE FALSE   TRUE   FALSE    TRUE       <NA>
#> 6        16   FALSE    FALSE  TRUE FALSE  FALSE   FALSE   FALSE       <NA>
#> 7        16   FALSE    FALSE FALSE FALSE  FALSE   FALSE   FALSE       <NA>
#> 8         2   FALSE    FALSE FALSE  TRUE  FALSE   FALSE   FALSE       <NA>
#> 9         2   FALSE    FALSE FALSE FALSE  FALSE   FALSE   FALSE       <NA>
#> 10       4.   FALSE    FALSE FALSE FALSE   TRUE    TRUE   FALSE       <NA>
#> 11        8    TRUE    FALSE FALSE FALSE  FALSE   FALSE   FALSE       <NA>
#> 12        4    TRUE     TRUE FALSE FALSE  FALSE   FALSE   FALSE       <NA>
#> 13        4   FALSE     TRUE FALSE FALSE  FALSE   FALSE   FALSE       <NA>
#> 14        8   FALSE    FALSE FALSE FALSE  FALSE   FALSE   FALSE       <NA>
#> 15        8   FALSE    FALSE FALSE FALSE  FALSE   FALSE   FALSE       <NA>
#> 16        8   FALSE    FALSE FALSE FALSE  FALSE   FALSE   FALSE       <NA>
#> 17        8   FALSE    FALSE FALSE FALSE  FALSE   FALSE   FALSE       <NA>
#> 18        1   FALSE    FALSE FALSE FALSE  FALSE   FALSE   FALSE       <NA>
#> 19        1   FALSE    FALSE FALSE FALSE  FALSE   FALSE   FALSE       <NA>
#>    articulation
#> 1          <NA>
#> 2          <NA>
#> 3            -.
#> 4          <NA>
#> 5          <NA>
#> 6          <NA>
#> 7          <NA>
#> 8          <NA>
#> 9          <NA>
#> 10         <NA>
#> 11         <NA>
#> 12         <NA>
#> 13         <NA>
#> 14         <NA>
#> 15         <NA>
#> 16         <NA>
#> 17         <NA>
#> 18         <NA>
#> 19         <NA>