Create a music track

Create a music track from a collection of musical phrases.

track(phrase, tuning = "e, a, d g b e'", add_staff = "treble_8",
  voice = 1L)

Arguments

phrase
a phrase object.

tuning
character, space-delimited pitches describing the guitar tuning. Defaults to standard tuning. Tick or integer octave numbering accepted.

add_staff
add a standard sheet music staff above the tablature staff. See details.

voice
integer, ID indicating phrases belonging to unique voices within a single track. May be a vector.

Value

a track table.

Details

Musical phrases generated by phrase are fortified in a track table. All tracks are stored as track tables, one per row, even if that table consists of a single track. track creates a single-entry track table. See trackbind for merging single tracks into a multi-track table. This is simply row binding that properly preserves phrase and track classes.

The default for an additional staff is add_staff = "treble_8" for 8va treble clef, which is commonly displayed in quality guitar tablature above the tablature staff to include precise rhythm and timing information. Note that guitar is a transposing instrument. For this reason, the default is "treble_8", not "treble". Set add_staff = NA to suppress the additional staff. This is appropriate for simple guitar patterns where there are already multiple guitar tracks and the additional space required for two staves per instrument is unnecessary and wasteful.

Examples

x <- phrase("c ec'g' ec'g'", "4 4 2", "5 432 432")
track(x)

#> # A tibble: 1 x 4
#>   phrase       staff    tuning         voice
#>   <list>       <chr>    <chr>          <int>
#> 1 <S3: phrase> treble_8 e, a, d g b e'     1