Bind track tables

Bind together track tables by row.

trackbind(...)

Arguments

...
track tables.

Value

a track table.

Details

This function appends multiple track tables into a single track table for preparation of generating a multi-track score. Regardless of whether a tabstaff ID column is present in the input tracks, one is added or the existing one is overridden ensuring a unique integer ID for each track. This is used to separate music staves in the sheet music/tablature output. The actual ID values assigned to each track do not matter; only the order in which tracks are bound, first to last.

Examples

x <- phrase("c ec'g' ec'g'", "4 4 2", "5 432 432")
trackbind(track(x), track(x))

#> # A tibble: 2 x 4
#>   phrase       staff    voice tabstaff
#>   <list>       <chr>    <int>    <int>
#> 1 <S3: phrase> treble_8     1        1
#> 2 <S3: phrase> treble_8     1        2