Transpose pitch

Transpose pitch by a number of semitones.

transpose(notes, n = 0, key = NULL)

tp(...)

Arguments

notes
character, a valid string of notes, the type passed to phrase.

n
integer, positive or negative number of semitones to transpose.

key
character, the new key signature after transposing notes. If provided, this helps ensure proper use of sharps vs. flats.

...
arguments passed to transpose.

Value

a character string.

Details

This function transposes the pitch of notes in a valid character string. The string must be of the form passed to the info argument to phrase.

Transposing is not done on a phrase object. The notes in a phrase object have already been transformed to LilyPond syntax and mixed with other potentially complex and variable information. Transposing is intended to be done on a string of notes prior to passing it to phrase. It will work on strings that use either integer or tick mark octave numbering formats. The transposed result will be a string with integer octave numbering.

The only element other pitch that occurs in a valid notes string is a rest, "r". This is ignored by transpose.

Examples

transpose("a_3 b_4 c5", 0)

#> [1] "a_3 b_4 c5"

tp("a_3 b_4 c5", -1)

#> [1] "g3 a4 b4"

tp("a_3 b_4 c5", 1)

#> [1] "a3 b4 c#5"

tp("a#3 b4 c#5", 11)

#> [1] "a4 a#5 c6"

tp("a#3 b4 c#5", 12)

#> [1] "a#4 b5 c#6"

tp("a#3 b4 c#5", 13)

#> [1] "b4 c6 d6"

tp("a3 b4 c5", 2, key = "f")

#> [1] "b3 d_5 d5"

tp("a3 b4 c5", 2, key = "g")

#> [1] "b3 c#5 d5"

tp("a b' c''", 2, key = "f")

#> [1] "b3 d_5 d5"

tp("a, b c'", 2, key = "g")

#> [1] "b2 c#4 d4"