Check note and chord validity

Check whether a string is comprised exclusively of valid note and/or chord substring syntax. is_note and is_chord are vectorized and their positive results are mutually exclusive. noteworthy is also vectorized and performs both checks, but it returns a scalar logical result indicating whether the entire set contains exclusively valid entries.

is_note(x)

is_chord(x)

noteworthy(x)

is_diatonic(x, key = "c")

as_noteworthy(x)

Arguments

x
character, space-delimited entries or a vector of single, non-delimited entries.

key
character, key signature.

Value

logical

Details

as_noteworthy can be used to coerce to the noteworthy class. Coercion will fail if the string is not noteworthy. Using the noteworthy class is generally not needed by the user during an interactive session, but is available and offers its own print and summary methods for noteworthy strings. It is more likely to be used by other functions and functions that output a noteworthy string generally attach the noteworthy class.

is_diatonic performs a vectorized logical check on a noteworthy string for all notes and chords. To check strictly notes or chords, see note_in_scale and chord_is_diatonic.

See also

Examples

x <- "a# b_ c, d'' e3 g_4 A m c2e_2g2 cegh"
data.frame(
  x = strsplit(x, " ")[[1]],
  note = is_note(x),
  chord = is_chord(x),
  either = noteworthy(x))

#>          x  note chord either
#> 1       a#  TRUE FALSE  FALSE
#> 2       b_  TRUE FALSE  FALSE
#> 3       c,  TRUE FALSE  FALSE
#> 4      d''  TRUE FALSE  FALSE
#> 5       e3  TRUE FALSE  FALSE
#> 6      g_4  TRUE FALSE  FALSE
#> 7        A FALSE FALSE  FALSE
#> 8        m FALSE FALSE  FALSE
#> 9  c2e_2g2 FALSE  TRUE  FALSE
#> 10    cegh FALSE FALSE  FALSE

is_diatonic("ace ac#e d e_", "c")

#> [1]  TRUE FALSE  TRUE FALSE

x <- "a# b_ c,~ c, d'' e3 g_4 c2e_2g2"
x <- as_noteworthy(x)
x

#> <Noteworthy string>
#>   Format: space-delimited time
#>   Values: a# b_ c,~ c, d'' e3 g_4 <c2e_2g2>

summary(x)

#> <Noteworthy string>
#>   Timesteps: 8 (7 notes, 1 chord)
#>   Octaves: ambiguous
#>   Accidentals: both/ambiguous
#>   Format: space-delimited time
#>   Values: a# b_ c,~ c, d'' e3 g_4 <c2e_2g2>